característica: implementación de compuertas (Gates)

+++ app/Http/Controllers/RoleController.php
@@ -12,22 +12,27 @@ use App\Http\Requests\UpRoleRequest;
 use Illuminate\Http\RedirectResponse;
 use Illuminate\View\View;
 
+
 class RoleController extends Controller
 {
     public function index(): View
     {
+        $this->authorize('haveAccess', 'role.index');
+
         $roles = Role::orderBy('id', 'Desc')->paginate(5);
         return view('role.index', compact('roles'));
     }
 
     public function create(): View
     {
+        $this->authorize('haveAccess', 'role.create');
         $permissions = Permission::get();
         return view('role.create', compact('permissions'));
     }
 
     public function store(StoreRoleRequest $request, Role $role): RedirectResponse
     {
+        $this->authorize('haveAccess', 'role.create');
         $role = Role::create($request->validated());
         StoreOrUpdateRole::execute($request, $role);
         return redirect()->route('role.index')
@@ -36,6 +41,7 @@ class RoleController extends Controller
 
     public function show(Role $role): View
     {
+        $this->authorize('haveAccess', 'role.show');
         $permissions_role = ShowOrEditRole::execute($role);
         $permissions = Permission::get();
         return view('role.show', compact('permissions', 'role', 'permissions_role'));
@@ -43,6 +49,7 @@ class RoleController extends Controller
 
     public function edit(Role $role): View
     {
+        $this->authorize('haveAccess', 'role.edit');
         $permissions_role = ShowOrEditRole::execute($role);
         $permissions = Permission::get();
         return view('role.edit', compact('permissions', 'role', 'permissions_role'));
@@ -50,6 +57,7 @@ class RoleController extends Controller
 
     public function update(UpRoleRequest $request, Role $role)
     {
+        $this->authorize('haveAccess', 'role.edit');
         $role->fill($request->validated());
         StoreOrUpdateRole::execute($request, $role);
         return redirect()->route('role.index')
@@ -58,6 +66,7 @@ class RoleController extends Controller
 
     public function destroy(Role $role)
     {
+        $this->authorize('haveAccess', 'role.destroy');
         $role->delete();
         return redirect()->route('role.index')
             ->with('status_success', 'Rol eliminado correctamente');
